moving BitmapWordReader and BitmapWordWriter to header files
